# docs/swagger.yaml
definitions:
  model.Game:
    properties:
      created_at:
        $ref: '#/definitions/model.MySQLTime'
      description:
        type: string
      developer:
        type: string
      id:
        type: integer
      release_date:
        type: string
      title:
        type: string
      updated_at:
        $ref: '#/definitions/model.MySQLTime'
    type: object
  model.GameResponse:
    properties:
      title:
        type: string
    type: object
  model.MySQLTime:
    properties:
      time.Time:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /games:
    get:
      description: Get a list of game titles
      responses:
        "200":
          description: List of game titles" // Sesuaikan dengan tipe m.GameResponse
          schema:
            items:
              $ref: '#/definitions/model.GameResponse'
            type: array
        "500":
          description: Internal server error" (when there is a problem with the database)
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get games
    post:
      description: Add a new game with the provided information (admin only)
      parameters:
      - description: Game object that needs to be added
        in: body
        name: game
        required: true
        schema:
          $ref: '#/definitions/model.Game'
      responses:
        "200":
          description: Game added
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid Release Date format. Use 'YYYY-MM-DD'." (when the provided
            Release Date has an invalid format)
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized" (when the provided JWT token is invalid or missing)
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Access denied" (when the user does not have admin role)
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error" (when there is a problem with the database)
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Add new game
  /games/{id}:
    delete:
      description: Delete a game by its ID (admin only)
      parameters:
      - description: Game ID to be deleted
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Game successfully deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid game ID" (when the provided game ID in the URL is not
            a valid integer)
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized" (when the provided JWT token is invalid or missing)
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Access denied" (when the user does not have admin role)
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Game not found" (when the requested game ID does not exist
            in the database)
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error" (when there is a problem with the database)
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete game
    get:
      description: Get detailed information about a specific game
      parameters:
      - description: Game ID to be retrieved
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Game details
          schema:
            $ref: '#/definitions/model.Game'
        "400":
          description: Invalid game ID" (when the provided game ID in the URL is not
            a valid integer)
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Game not found" (when the requested game ID does not exist
            in the database)
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error" (when there is a problem with the database)
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get game details
    post:
      description: Update game data (title, developer, release date, and description)
        (admin only)
      parameters:
      - description: Game ID to be updated
        in: path
        name: id
        required: true
        type: integer
      - description: Game object that contains updated game data
        in: body
        name: game
        required: true
        schema:
          $ref: '#/definitions/model.Game'
      responses:
        "200":
          description: Game data updated
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid Release Date format. Use 'YYYY-MM-DD'." (when the provided
            Release Date has an invalid format)
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized" (when the provided JWT token is invalid or missing)
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Access denied" (when the user does not have admin role)
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Game not found" (when the requested game ID does not exist
            in the database)
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error" (when there is a problem with the database)
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update game
...
swagger: "2.0"